fixing up targets/props so we can build with -test
