+++ eng/WpfArcadeSdk/tools/DRT.Build.props
@@ -1,5 +1,7 @@
 <Project>
   <PropertyGroup>
     <PublishDir>$(WpfDrtTestBasePublishPath)</PublishDir>
+    <IsTestProject>true</IsTestProject>
+    <SkipTests>true</SkipTests>
   </PropertyGroup>
 </Project>
