+++ eng/WpfArcadeSdk/tools/StrongName.targets
