+++ eng/copy-wpf.ps1
@@ -60,7 +60,10 @@ function CopyPackagedBinaries($location, $localBinLocation, $packageName, $binar
 {
     $ArchFolder = if ($arch -eq "x86") { "" } else { "x64" }
     $BinLocation = [System.IO.Path]::Combine($localBinLocation, $Config, $ArchFolder, $packageName, "lib", $binaryLocationInPackage, "*")
-    Copy-Item -path $BinLocation -include "*.dll","*.pdb" -Destination $location
+    if (Test-Path $BinLocation)
+    {
+        Copy-Item -path $BinLocation -include "*.dll","*.pdb" -Destination $location
+    }
 }
 
 if ($help -or ([string]::IsNullOrEmpty($destination) -and !$local))
