+++ eng/helix/prepare-helix-payload.ps1
@@ -1,7 +1,8 @@
 [CmdLetBinding()]
 Param(
     [string]$platform,
-    [string]$configuration
+    [string]$configuration,
+    [string]$testPackageVersion
 )
 
 $payloadDir = "HelixPayload\$configuration\$platform"
@@ -34,10 +35,11 @@ function CopyFolderStructure($from, $to)
         Write-Output "Location doesn't exist: $from"
     }
 
+    Get-ChildItem $to -Recurse
 }
 
-# Copy files from nuget packages
-$testNugetLocation = Join-Path $nugetPackagesDir "runtime.win-$platform.Microsoft.DotNet.Wpf.DncEng.Test\1.0.0-beta.19260.5\tools\win-$platform\"
+# Copy files from nuget packages.
+$testNugetLocation = Resolve-Path (Join-Path $nugetPackagesDir "runtime.win-$platform.Microsoft.DotNet.Wpf.Test\*\tools\win-$platform\Test")
 $testPayloadLocation = Join-Path $payloadDir "Test"
 CopyFolderStructure $testNugetLocation $testPayloadLocation
 
@@ -46,16 +48,6 @@ CopyFolderStructure $testNugetLocation $testPayloadLocation
 $drtArtifactsLocation = [System.IO.Path]::Combine($env:BUILD_SOURCESDIRECTORY, "artifacts\test", $configuration, $platform, "Test\DRT")
 $drtPayloadLocation = Join-Path $payloadDir "Test\DRT"
 CopyFolderStructure $drtArtifactsLocation $drtPayloadLocation
-
-# Copy built assemblies to dotnet install location
-$eng = Join-Path $env:BUILD_SOURCESDIRECTORY "eng"
-$configArgs = if ($configuration == "Release") { "-release" } else { }
-& "$eng\copy-wpf.ps1 -local -arch $platform $configArgs"
-
-# Copy local dotnet install to payload
-$localDotnetInstall = Join-Path $env:BUILD_SOURCESDIRECTORY ".dotnet"
-$dotnetPayloadLocation = Join-Path $payloadDir "dotnet"
-CopyFolderStructure $localDotnetInstall $dotnetPayloadLocation
 
 # Copy scripts
 Copy-Item "eng\helix\configure-helix-machine.ps1" $payloadDir
