+++ src/Microsoft.DotNet.Wpf/test/Common/DRT/TestServices/DrtBaseGlobalInput.cs
