+++ src/Microsoft.DotNet.Wpf/test/Common/DRT/TestServices/DrtBaseInput.cs
