+++ src/Microsoft.DotNet.Wpf/test/Common/DRT/TestServices/MS/Win32/HandleCollector.cs
