+++ src/Microsoft.DotNet.Wpf/test/DRT/DrtXaml/FriendWithKey/XamlTestClasses.FriendWithKey.csproj
@@ -3,7 +3,6 @@
     <AssemblyName>XamlTestClasses.FriendWithKey</AssemblyName>
     <OutputType>Library</OutputType>
     <DefineConstants>$(DefineConstants);OLDRESOURCES</DefineConstants>
-    <TestCode>true</TestCode>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
   <ItemGroup>
