+++ src/Microsoft.DotNet.Wpf/test/DRT/DrtXaml/PartialTrustTests/DrtXaml.PartialTrustTests.csproj
@@ -3,7 +3,6 @@
     <AssemblyName>DrtXaml.PartialTrustTests</AssemblyName>
     <OutputType>Library</OutputType>
     <DefineConstants>$(DefineConstants);OLDRESOURCES</DefineConstants>
-    <TestCode>true</TestCode>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <AssemblyAttributeClsCompliant>false</AssemblyAttributeClsCompliant>
   </PropertyGroup>
