--- src/Microsoft.DotNet.Wpf/test/PublishAfterBuild.targets
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project>
-   <Target Name="AfterBuild" DependsOnTargets="PublishPropertySetup;Publish"/> 
-  
-   <Target
-       Name="PublishPropertySetup">
-     <PropertyGroup>
-         <NoBuild>true</NoBuild>
-     </PropertyGroup>
-   </Target>
-</Project>
